Work in progress for js interface.
eWay & Authorize.net are similar in that a script needs to be included & it will transform fields.

The transformed version can be submitted to the same php handling but we
probably need to define the 'client_side_encrypted' fields in the metadata. That way we can check
non-emptiness of them as being isClientSideEncrypted and that will affect validation (not
the current hack) and we can also determine vars  - probably for getCreditCardObjectParams.

I think the change in eway code currently will be reverted for the metadata above.

Note that I found the eway js would not run unless I removed from the core tpl the
click return='SubmitOnce' code. Not sure why. Perhaps that should be submit not click?